[Feat] #39 - VoiceListenerView 분리
